Move gnocchi.storage.incoming to gnocchi.incoming
