+++ gnocchi/tests/base.py
@@ -33,6 +33,7 @@ from tooz import coordination
 
 from gnocchi import archive_policy
 from gnocchi import exceptions
+from gnocchi import incoming
 from gnocchi import indexer
 from gnocchi import service
 from gnocchi import storage
@@ -348,7 +349,7 @@ class TestCase(BaseTestCase):
                                "storage")
 
         self.storage = storage.get_driver(self.conf)
-        self.incoming = storage.get_incoming_driver(self.conf)
+        self.incoming = incoming.get_driver(self.conf)
 
         if self.conf.storage.driver == 'redis':
             # Create one prefix per test
