+++ setup.cfg
@@ -109,11 +109,11 @@ gnocchi.storage =
     redis = gnocchi.storage.redis:RedisStorage
 
 gnocchi.incoming =
-    ceph = gnocchi.storage.incoming.ceph:CephStorage
-    file = gnocchi.storage.incoming.file:FileStorage
-    swift = gnocchi.storage.incoming.swift:SwiftStorage
-    s3 = gnocchi.storage.incoming.s3:S3Storage
-    redis = gnocchi.storage.incoming.redis:RedisStorage
+    ceph = gnocchi.incoming.ceph:CephStorage
+    file = gnocchi.incoming.file:FileStorage
+    swift = gnocchi.incoming.swift:SwiftStorage
+    s3 = gnocchi.incoming.s3:S3Storage
+    redis = gnocchi.incoming.redis:RedisStorage
 
 gnocchi.indexer =
     mysql = gnocchi.indexer.sqlalchemy:SQLAlchemyIndexer
